+++ communication/communicators/comm_t.hpp
@@ -19,7 +19,7 @@ comm_t* ini_comm(const char* name, DIRECTION dir, const COMM_TYPE &t);
 //int comm_send(comm_t* comm, const char *data, const size_t &len);
 //int comm_recv(comm_t* comm, char **data, const size_t &len, bool allow_realloc);
 int comm_send(comm_t* comm, const dtype_t* dtype);
-int comm_recv(comm_t* comm, dtype_t* dtype);
+long comm_recv(comm_t* comm, dtype_t* dtype);
 int comm_nmsg(comm_t* comm);
 #ifdef __cplusplus
 }
